+++ src/assets/js/chart/chart.js
@@ -3,11 +3,15 @@ import nav from './nav';
 import legend from './legend';
 import xAxis from './xAxis';
 import primary from './primary';
+import multiChart from './multiChart';
+import group from './group';
 
 export default {
     legend: legend,
     nav: nav,
     primary: primary,
     xAxis: xAxis,
-    secondary: secondary
+    secondary: secondary,
+    multiChart: multiChart,
+    group: group
 };