+++ src/assets/js/model/chart/chart.js
@@ -1,3 +1,4 @@
+import group from './group';
 import legend from './legend';
 import nav from './nav';
 import navigationReset from './navigationReset';
@@ -6,6 +7,7 @@ import secondary from './secondary';
 import xAxis from './xAxis';
 
 export default {
+    group: group,
     legend: legend,
     nav: nav,
     navigationReset: navigationReset,