+++ src/assets/js/model/chart/secondary.js
@@ -1,9 +1,11 @@
 export default function(initialProduct, initialDiscontinuityProvider) {
     return {
         data: [],
+        visibleData: [],
         viewDomain: [],
         trackingLatest: true,
         product: initialProduct,
-        discontinuityProvider: initialDiscontinuityProvider
+        discontinuityProvider: initialDiscontinuityProvider,
+        indicators: []
     };
 }