+++ src/assets/js/util/layout.js
@@ -1,32 +1,10 @@
-/*global window */
 import fc from 'd3fc';
 
 var renderedOnce = false;
 
 export default function(containers, charts) {
-
-    function getSecondaryContainer(chartIndex) {
-        return containers.secondaries.filter(function(d, index) { return index === chartIndex; });
-    }
-
-    var secondaryChartsShown = 0;
-    for (var j = 0; j < charts.secondaries.length; j++) {
-        if (charts.secondaries[j]) {
-            secondaryChartsShown++;
-        }
-    }
-    containers.secondaries
-        .filter(function(d, index) { return index < secondaryChartsShown; })
-        .style('flex', '1');
-    containers.secondaries
-        .filter(function(d, index) { return index >= secondaryChartsShown; })
-        .style('flex', '0');
-    containers.overlaySecondaries
-        .filter(function(d, index) { return index < secondaryChartsShown; })
-        .style('flex', '1');
-    containers.overlaySecondaries
-        .filter(function(d, index) { return index >= secondaryChartsShown; })
-        .style('flex', '0');
+    containers.secondaries.style('flex', charts.secondaries().charts().length);
+    containers.overlaySecondaries.style('flex', charts.secondaries().charts().length);
 
     var headRowHeight = parseInt(containers.app.select('.head-row').style('height'), 10);
     if (!renderedOnce) {
@@ -41,10 +19,10 @@ export default function(containers, charts) {
 
     containers.chartsAndOverlay.style('height', useableHeight + 'px');
 
-    charts.xAxis.dimensionChanged(containers.xAxis);
-    charts.navbar.dimensionChanged(containers.navbar);
-    charts.primary.dimensionChanged(containers.primary);
-    for (var i = 0; i < charts.secondaries.length; i++) {
-        charts.secondaries[i].option.dimensionChanged(getSecondaryContainer(i));
-    }
+    charts.xAxis().dimensionChanged(containers.xAxis);
+    charts.nav().dimensionChanged(containers.navbar);
+    charts.primary().dimensionChanged(containers.primary);
+    charts.secondaries().charts().forEach(function(chart) {
+        chart.option.dimensionChanged(containers.secondaries);
+    });
 }